testPython/requirements.txt
behave
kafka-python
sqlalchemy
psycopg2-binary
requests
